docs(typedoc): Use generateReferenceDocs function
