+++ package.json
@@ -27,7 +27,7 @@
     "dev": "pnpm run watch",
     "prettier": "prettier --ignore-unknown '**/*'",
     "prettier:write": "pnpm run prettier --write",
-    "docs:reference": "node scripts/reference-docs.js",
+    "docs:generate": "node scripts/generateDocs.js",
     "cipublish": "node scripts/publish.js"
   },
   "nx": {
@@ -42,7 +42,7 @@
     "@cspell/eslint-plugin": "^8.9.1",
     "@eslint-react/eslint-plugin": "^1.5.16",
     "@solidjs/testing-library": "^0.8.8",
-    "@tanstack/config": "^0.9.8",
+    "@tanstack/config": "^0.10.0",
     "@testing-library/jest-dom": "^6.4.6",
     "@testing-library/react": "^16.0.0",
     "@types/eslint": "^8.56.10",
@@ -66,9 +66,6 @@
     "sherif": "^0.10.0",
     "solid-js": "^1.8.17",
     "tsup": "^8.0.2",
-    "typedoc": "^0.26.4",
-    "typedoc-plugin-frontmatter": "^1.0.0",
-    "typedoc-plugin-markdown": "^4.2.1",
     "typescript": "5.3.3",
     "typescript47": "npm:typescript@4.7",
     "typescript48": "npm:typescript@4.8",
